+++ src/routes/jira/jira-router.ts
@@ -7,7 +7,6 @@ import { JiraSelectVersionRouter } from "./server/jira-select-version-router";
 import { JiraContextJwtTokenMiddleware } from "middleware/jira-jwt-middleware";
 import { JiraServerUrlRouter } from "routes/jira/server/jira-server-url-router";
 import { JiraAppCreationRouter } from "./server/jira-app-creation-router";
-import { JiraManualAppCreationRouter } from "./server/jira-manual-app-creation-router";
 
 export const JiraRouter = Router();
 
@@ -19,4 +18,3 @@ JiraRouter.use("/events", JiraEventsRouter);
 JiraRouter.use("/select-version", JiraSelectVersionRouter);
 JiraRouter.use("/server-url", JiraServerUrlRouter);
 JiraRouter.use("/app-creation", JiraAppCreationRouter);
-JiraRouter.use("/manual-app-creation", JiraManualAppCreationRouter);
