+++ views/jira-select-github-version.hbs
@@ -71,7 +71,6 @@
     </div>
 
     <script src="/public/js/jira-select-card-option.js" nonce="{{nonce}}"></script>
-    <script src="/public/js/jira-select-github-version.js" nonce="{{nonce}}"></script>
     <script src="/public/js/navigation.js" nonce="{{nonce}}"></script>
     <!-- Per https://blog.developer.atlassian.com/announcement-reminder-about-deprecation-of-xdm_e-usage-and-needing-to-load-all-js-from-the-cdn/ we are required to load this from this specific CDN -->
     <!-- DO NOT TOUCH!!! THIS IS NEEDED FOR CONNECT OR ELSE IT WILL CAUSE AN ERROR -->
